% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomSupportPkg-package.R
\docType{package}
\name{CustomSupportPkg}
\alias{CustomSupportPkg}
\alias{CustomSupportPkg-package}
\title{An Example Shiny App Support Package}
\description{
An Example Shiny App Support Package
}
\author{
You (you@exampole.com)
}

